Use curly braces for all variables
Curly braces aid the interpreter to identify which part of a composed
string is the variable part to expand, and what the string to
concatenate to is.
In the case of
```bash
${volume_name}-${date_suffix}.tar
```
it worked, because -, and . are not allowd in variable names.

Use xz compression
Add proper file extension to tar invocations